Props and Textures
I have added all the props and textures to the scene, I have also fixed the lighting issue. In the plug was much bigger when scale factor was turned off so I have made a varient of it which correctly sized the fit on one block. I scale the hight to be a bit bigger though